Small fix to prevent annoying warning messages in the log.
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@784552 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/manufacturing/src/org/ofbiz/manufacturing/mrp/InventoryEventPlannedServices.java
@@ -66,7 +66,7 @@ public class InventoryEventPlannedServices {
         mrpEvent = delegator.findByPrimaryKey("MrpEvent", mrpEventKeyMap);
         if (mrpEvent == null) {
             mrpEvent = delegator.makeValue("MrpEvent", mrpEventKeyMap);
-            mrpEvent.put("quantity", newQuantity);
+            mrpEvent.put("quantity", newQuantity.doubleValue());
             mrpEvent.put("eventName", eventName);
             mrpEvent.put("facilityId", facilityId);
             mrpEvent.put("isLate", (isLate? "Y": "N"));
